Show single calendar only on small devices
In a previous commit, I fixed the display of dual calendars on small
devices: they appeared one on top of the other. This avoided the
_horizontal_ scroll... but actually now the use has to scroll
_vertically_ to see the bottom of the second calendar.

We don't want scrolling, so let's simplify:
- on wide screens (that supposedly have a large height): show dual
  calendars;
- on narrow screens (that supposedly have a small height): show only
  one calendar.

The code is not pretty (for example, some HTML is duplicated) but
we can surely live with it for a while.